Fix some Eclipse warnings
git-svn-id: https://svn-us.apache.org/repos/asf/commons/proper/jexl/trunk@889768 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/jexl2/ClassCreator.java
@@ -128,7 +128,7 @@ public class ClassCreator {
             return null;
         }
         Integer r = (Integer) jexl.invokeMethod(javac, "compile", source);
-        if (r >= 0) {
+        if (r.intValue() >= 0) {
             return getClassLoader().loadClass("org.apache.commons.jexl2.generated." + className);
         }
         return null;
